feat: add Java REST API in the back end (#105)
* feat: add backend initial files

Issue #91

This commit adds initial files generated by start.spring.io website
where the project was generated with:

- Project: Maven
- Language: Java
- Spring Boot version: 3.1.3 (latest stable)
- Project metadata group: ca.bc.gov.restapi
- Project metadata artifact: results
- Project metadata name: results
- Project metadata description: RESULTS REST API
- Packaging: Jar
- Java version: 17
- Dependencies
  - Actuator
  - Data JPA
  - OAuth2 Authorization server
  - OAuth2 client
  - Started web
  - Devtools
  - H2 Database
  - Lombok
  - GraalVM

* feat: add maven wrapper jar

Issue #91

Adding the maven wrapper is a good practice, recommended in case you
want to use maven out of the box.

* feat: add plugins and review pom.xml backend file

Issue #91

This change adds required dependencies for building, running and
packaging the service locally and for Cloud Native images. This is what
was changed:

- Add project license
- Add properties and definitions
- Add profiles for dev, prod and native
- Add dependencies
- Add plugins for building, running and packaging
- Add project final name jar, line 179

* feat: add google checkstyle

Issue #91

This commit adds the google checkstyle xml definition file to be used as
a guide for the checkings. Also updates existing class files to met the
checkstyle validations, replacing tab by spaces, adding JavaDoc to all
public classes.

* feat: add backend to docker compose file

Issue #91

This change adds a new service to the docker-compose file for the
backend REST api. If you want to run the backend locally all you need is
to run at the project root folder: `docker compose up backend` and
you're all set.

Other minor changes were made:
- Add some properties to the application properties file
- Remove dependencies not required for now

* test: fix test cases and tests configuration

Issue #91

This change gets tests passing and cloud native build working and
running smoothly.

* docs: update readme file to include new service

Issue #91

This change simply add more information on the new service in the back
end.

* feat: update sb version to 3.1.4 and java to 21

Issue #91

This change updates the Spring Boot version to the latest (at the time
of this writing), which is 3.1.4. And also updates the required Java
version required to compile and run this service, which is the 21.

* feat: update docker compose to run with java 21

Issue #91

This commit updates the existing docker-compose file to use the latest
maven with Java 21 available at this moment.

* fix: trivy warning cve-2022-1471

Issue #91

This change simply change the required version by spring boot of this
dependency. You can learn more about it here:

- https://github.com/spring-projects/spring-boot/issues/32221
- https://avd.aquasec.com/nvd/2022/cve-2022-1471/